+++ CHANGELOG.md
@@ -13,6 +13,8 @@ is based on [Keep a Changelog](https://keepachangelog.com).
 ### Added
 
 - New flow operator: `retry`.
+- New `with_userinfo` member function for URIs that allows setting the user-info
+  sub-component without going through an URI builder.
 
 ### Fixed